+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: yallo <yallo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 15:48:33 by yallo             #+#    #+#             */
-/*   Updated: 2022/11/09 14:29:16 by yallo            ###   ########.fr       */
+/*   Updated: 2022/11/14 13:25:56 by yallo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
